+++ src/main/java/blackjack/domain/gamer/Bet.java
@@ -4,9 +4,9 @@ import java.util.Objects;
 
 public class Bet {
 
-	public static final String INVALID_RANGE_ERROR = "금액은 0보다 커야 합니다";
-	public static final int INVALID_MAX_VALUE = 0;
-	public static final String INAVLID_DEVIDIBLE_VALUE_ERROR = "금액은 10원 단위여야 합니다.";
+	private static final String INVALID_RANGE_ERROR = "금액은 0보다 커야 합니다";
+	private static final int INVALID_MAX_VALUE = 0;
+	private static final String INAVLID_DEVIDIBLE_VALUE_ERROR = "금액은 10원 단위여야 합니다.";
 
 	private final int amount;
 
@@ -25,8 +25,12 @@ public class Bet {
 		}
 	}
 
-	public int getAmount() {
-		return amount;
+	public static int getMinusAmount(int amount) {
+		return amount * -1;
+	}
+
+	public int multiply(double value) {
+		return (int) (value * amount);
 	}
 
 	@Override
