+++ src/test/java/blackjack/domain/result/BlackJackResultTest.java
@@ -185,15 +185,6 @@ class BlackJackResultTest {
 		assertThat(result.getProfit()).isEqualTo(0);
 	}
 
-	@ParameterizedTest
-	@CsvSource(value = {"BLACKJACK_WIN:1500", "WIN:1000", "LOSE:-1000", "DRAW:0"}, delimiter = ':')
-	@DisplayName("수익률에 따른 수익 계산")
-	void calculateEarning(String input, int bet) {
-		BlackJackResult result = BlackJackResult.valueOf(input);
-		int earning = result.calculateEarning(1000);
-		assertThat(earning).isEqualTo(bet);
-	}
-
 	private Card getAce() {
 		return Card.getInstance(CardShape.CLOVER, CardNumber.ACE);
 	}
